% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{ex_cf}
\alias{ex_cf}
\title{Uses the extract_value function to extract the chemical formula from a CIF file}
\usage{
ex_cf(cif_content)
}
\arguments{
\item{cif_content}{a string containing the content of a CIF file}
}
\value{
The chemical formula of the compound
}
\description{
Uses the extract_value function to extract the chemical formula from a CIF file
}
\examples{
\dontrun{
extract_chemical_formula(cif_content)
}
}
