% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{ex_sg_num}
\alias{ex_sg_num}
\title{Uses the extract_value function to extract the space group number from a CIF file}
\usage{
ex_sg_num(cif_content)
}
\arguments{
\item{cif_content}{a string containing the content of a CIF file}
}
\value{
The space group number
}
\description{
Uses the extract_value function to extract the space group number from a CIF file
}
\examples{
\dontrun{
extract_space_group_number(cif_content)
}
}
