% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{ex_st}
\alias{ex_st}
\title{Uses the extract_value function to extract the structure type from a CIF file}
\usage{
ex_st(cif_content)
}
\arguments{
\item{cif_content}{a string containing the content of a CIF file}
}
\value{
The structure type of the compound
}
\description{
Uses the extract_value function to extract the structure type from a CIF file
}
\examples{
\dontrun{
extract_structure_type(cif_content)
}
}
