% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_ucm.R
\name{ex_ucm}
\alias{ex_ucm}
\title{Function to extract unit cell metrics}
\usage{
ex_ucm(cif_content)
}
\arguments{
\item{cif_content}{a string containing the content of a CIF file}
}
\value{
A data table with the extracted unit cell metrics
}
\description{
Function to extract unit cell metrics
}
\examples{
\dontrun{
extract_unit_cell_metrics(cif_content)
}
}
