% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_ac.R
\name{extract_atomic_coordinates}
\alias{extract_atomic_coordinates}
\title{Extracting atomic coordinates}
\usage{
extract_atomic_coordinates(cif_content)
}
\arguments{
\item{cif_content}{a string containing the content of a CIF file}
}
\value{
A data table with the extracted atomic coordinates
}
\description{
Extracting atomic coordinates
}
