% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cif_data.R
\name{extract_chemical_formula}
\alias{extract_chemical_formula}
\title{Extract Chemical Formula Sum from CIF Content}
\usage{
extract_chemical_formula(cif_content)
}
\arguments{
\item{cif_content}{A data.table representing the CIF file content.}
}
\value{
A character string with the chemical formula sum, or NA if not found.
}
\description{
Extract Chemical Formula Sum from CIF Content
}
\examples{
# cif_data <- data.table(V1 = c("_chemical_formula_sum 'H2 O'"))
# extract_chemical_formula(cif_data)
}
