% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cif_data.R
\name{extract_cif}
\alias{extract_cif}
\alias{extract_database_code}
\alias{extract_chemical_formula}
\alias{extract_structure_type}
\alias{extract_space_group_name}
\alias{extract_space_group_number}
\title{Extract Crystallographic Information from CIF Content}
\usage{
extract_database_code(cif_content)

extract_chemical_formula(cif_content)

extract_structure_type(cif_content)

extract_space_group_name(cif_content)

extract_space_group_number(cif_content)
}
\arguments{
\item{cif_content}{A data.table representing the CIF file content, typically from \code{fread(..., sep = "\\n", header = FALSE)}.}
}
\value{
A character string with the requested value, or NA if not found.
}
\description{
A suite of functions to extract specific metadata tags from CIF file content.
}
\examples{
# cif_data <- data.table(V1 = c("data_example", "_database_code_ '12345'"))
# extract_database_code(cif_data)
# cif_data <- data.table(V1 = c("_chemical_formula_sum 'H2 O'"))
# extract_chemical_formula(cif_data)
# cif_data <- data.table(V1 = c("_chemical_name_structure_type 'NaCl'"))
# extract_structure_type(cif_data)
# cif_data <- data.table(V1 = c("_space_group_name_H-M_alt 'P 1'"))
# extract_space_group_name(cif_data)
# cif_data <- data.table(V1 = c("_space_group_IT_number 1"))
# extract_space_group_number(cif_data)
}
