% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coordinates.R
\name{extract_coordinate_data}
\alias{extract_coordinate_data}
\alias{extract_atomic_coordinates}
\alias{extract_symmetry_operations}
\title{Extract Coordinate and Symmetry Data from CIF Content}
\usage{
extract_atomic_coordinates(cif_content)

extract_symmetry_operations(cif_content)
}
\arguments{
\item{cif_content}{A data.table representing the CIF file content, typically
from \code{fread(..., sep = "\\n", header = FALSE)}.}
}
\value{
A data.table with the extracted information, or NULL if not found.
}
\description{
Functions to extract atomic coordinates and symmetry operations from CIF file content.
}
\examples{
# cif_lines <- c("loop_", "_atom_site_label", "_atom_site_fract_x",
#                "_atom_site_occupancy",
#                "Si1 X X X 0.0 0.0 0.0 1.0")
# cif_data <- data.table(V1 = cif_lines)
# extract_atomic_coordinates(cif_data)
# cif_lines <- c("_space_group_symop_operation_xyz", "1 'x, y, z'", "2 '-x, -y, -z'")
# cif_data <- data.table(V1 = cif_lines)
# extract_symmetry_operations(cif_data)
}
