% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_ICSD_code}
\alias{extract_ICSD_code}
\title{Uses the extract_value function to extract the ICSD code from a CIF file}
\usage{
extract_ICSD_code(cif_content)
}
\arguments{
\item{cif_content}{a string containing the content of a CIF file}
}
\value{
The code from the ICSD database
}
\description{
Uses the extract_value function to extract the ICSD code from a CIF file
}
\examples{
\dontrun{
extract_ICSD_code(cif_content)
}
}
