% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cif_data.R
\name{extract_space_group_name}
\alias{extract_space_group_name}
\title{Extract Space Group Name (H-M) from CIF Content}
\usage{
extract_space_group_name(cif_content)
}
\arguments{
\item{cif_content}{A data.table representing the CIF file content.}
}
\value{
A character string with the space group name, or NA if not found.
}
\description{
Extract Space Group Name (H-M) from CIF Content
}
\examples{
# cif_data <- data.table(V1 = c("_space_group_name_H-M_alt 'P 1'"))
# extract_space_group_name(cif_data)
}
