% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cif_data.R
\name{extract_space_group_number}
\alias{extract_space_group_number}
\title{Extract Space Group IT Number from CIF Content}
\usage{
extract_space_group_number(cif_content)
}
\arguments{
\item{cif_content}{A data.table representing the CIF file content.}
}
\value{
A character string with the space group IT number, or NA if not found.
}
\description{
Extract Space Group IT Number from CIF Content
}
\examples{
# cif_data <- data.table(V1 = c("_space_group_IT_number 1"))
# extract_space_group_number(cif_data)
}
