% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cif_data.R
\name{extract_structure_type}
\alias{extract_structure_type}
\title{Extract Structure Type from CIF Content}
\usage{
extract_structure_type(cif_content)
}
\arguments{
\item{cif_content}{A data.table representing the CIF file content.}
}
\value{
A character string with the structure type, or NA if not found.
}
\description{
Extract Structure Type from CIF Content
}
\examples{
# cif_data <- data.table(V1 = c("_chemical_name_structure_type 'NaCl'"))
# extract_structure_type(cif_data)
}
