% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_value}
\alias{extract_value}
\title{Function to Extract a Value From a CIF File}
\usage{
extract_value(cif_content, pattern, remove_pattern = TRUE)
}
\arguments{
\item{cif_content}{a string containing the content of a CIF file}

\item{pattern}{a regular expression pattern to match the desired value}

\item{remove_pattern}{a logical value indicating whether to remove the pattern from the value}
}
\description{
Function to Extract a Value From a CIF File
}
\examples{
\dontrun{
 extract_value(cif_content, "data_ICSD_code")
 }
}
