% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World Function}
\usage{
hello(x)
}
\arguments{
\item{x}{A character string}
}
\value{
The output from \code{\link{print}}
}
\description{
Hello World Function
}
\examples{
hello("world")
\dontrun{
hello("Don")
}
}
