% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{minimum_distance}
\alias{minimum_distance}
\title{Minimum Distance Bonding}
\usage{
minimum_distance(distances, delta = 0.1)
}
\arguments{
\item{distances}{A data.table.}

\item{delta}{A numeric tolerance.}
}
\value{
A data.table.
}
\description{
Minimum Distance Bonding
}
\examples{
# See package vignette for examples.
}
