% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cif_files.R
\name{analyze_cif_files}
\alias{analyze_cif_files}
\title{Analyze Multiple CIF Files}
\usage{
analyze_cif_files(
  cif_input,
  pattern = "*.cif",
  bonding_method = "min_dist",
  ...
)
}
\arguments{
\item{cif_input}{Path to a folder or a vector of file paths.}

\item{pattern}{Glob pattern for files if `cif_input` is a folder.}

\item{bonding_method}{Bonding algorithm to use.}

\item{...}{Additional parameters passed to `process_single_cif_data`.}
}
\value{
A data.table where each row summarizes one CIF file.
}
\description{
Analyze Multiple CIF Files
}
