% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{calculate_neighbor_counts}
\alias{calculate_neighbor_counts}
\title{Calculate Neighbor Counts}
\usage{
calculate_neighbor_counts(bonded_pairs_table)
}
\arguments{
\item{bonded_pairs_table}{A `data.table` of bonded pairs.}
}
\value{
A `data.table` with columns 'Atom' and 'NeighborCount'.
}
\description{
Counts the number of nearest neighbors for each central atom based on
  a table of bonded pairs.
}
\seealso{
Other property calculators: 
\code{\link{calculate_angles}()},
\code{\link{calculate_distances}()}
}
\concept{property calculators}
