% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cif_data.R
\name{extract_chemical_formula}
\alias{extract_chemical_formula}
\title{Extract Chemical Formula from CIF Content}
\usage{
extract_chemical_formula(cif_content)
}
\arguments{
\item{cif_content}{A `data.table` where each row is a line from a CIF file.}
}
\value{
A character string of the chemical formula, or `NA` if not found.
}
\description{
Extract Chemical Formula from CIF Content
}
\seealso{
Other extractors: 
\code{\link{extract_atomic_coordinates}()},
\code{\link{extract_database_code}()},
\code{\link{extract_space_group_name}()},
\code{\link{extract_space_group_number}()},
\code{\link{extract_structure_type}()},
\code{\link{extract_symmetry_operations}()},
\code{\link{extract_unit_cell_metrics}()}
}
\concept{extractors}
