% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cif_files.R
\name{process_single_cif_data}
\alias{process_single_cif_data}
\title{Process a Single CIF File Content}
\usage{
process_single_cif_data(cif_content, bonding_method = "min_dist", ...)
}
\arguments{
\item{cif_content}{A data.table representing the content of one CIF file.}

\item{bonding_method}{A character string specifying the bonding algorithm ("min_dist", "brunner", "hoppe").}

\item{...}{Additional parameters passed to the bonding function.}
}
\value{
A data.table row containing all extracted and calculated data for the CIF file.
}
\description{
Process a Single CIF File Content
}
