% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cif_files.R
\name{read_cif_files}
\alias{read_cif_files}
\title{Read CIF Files from a List of Paths}
\usage{
read_cif_files(cif_file_paths)
}
\arguments{
\item{cif_file_paths}{A character vector of paths to CIF files.}
}
\value{
A list of data.tables.
}
\description{
Read CIF Files from a List of Paths
}
